Refactoring and separate
