+++ src/main/java/com/pakskiy/paymentProvider/dto/payment/PaymentResponseDto.java
@@ -2,18 +2,17 @@ package com.pakskiy.paymentProvider.dto.payment;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.pakskiy.paymentProvider.dto.TransactionStatus;
-import lombok.Builder;
+import com.pakskiy.paymentProvider.dto.TransactionResponseDto;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.experimental.SuperBuilder;
+
 
 @Data
-@Builder
+@SuperBuilder
+@EqualsAndHashCode(callSuper = true)
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class PaymentResponseDto {
+public class PaymentResponseDto extends TransactionResponseDto {
     @JsonProperty("transaction_id")
-    private final Long transactionId;
-    @JsonProperty("status")
-    private final TransactionStatus status;
-    @JsonProperty("message")
-    private String message;
+    private Long transactionId;
 }
