+++ src/main/java/com/pakskiy/paymentProvider/dto/payout/PayoutResponseDto.java
@@ -2,18 +2,16 @@ package com.pakskiy.paymentProvider.dto.payout;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.pakskiy.paymentProvider.dto.TransactionStatus;
-import lombok.Builder;
+import com.pakskiy.paymentProvider.dto.TransactionResponseDto;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.experimental.SuperBuilder;
 
 @Data
-@Builder
+@SuperBuilder
+@EqualsAndHashCode(callSuper = true)
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class PayoutResponseDto {
+public class PayoutResponseDto extends TransactionResponseDto {
     @JsonProperty("payout_id")
-    private Long payoutId;
-    @JsonProperty("status")
-    private TransactionStatus status;
-    @JsonProperty("message")
-    private String message;
+    private Long transactionId;
 }
