+++ src/main/java/com/pakskiy/paymentProvider/service/PayoutService.java
@@ -1,106 +1,17 @@
 package com.pakskiy.paymentProvider.service;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.pakskiy.paymentProvider.dto.payout.PayoutRequestDto;
 import com.pakskiy.paymentProvider.dto.payout.PayoutResponseDto;
-import com.pakskiy.paymentProvider.entity.MerchantEntity;
 import com.pakskiy.paymentProvider.entity.TransactionEntity;
-import com.pakskiy.paymentProvider.repository.CountryRepository;
-import com.pakskiy.paymentProvider.repository.CurrencyRepository;
-import com.pakskiy.paymentProvider.repository.LanguageRepository;
-import com.pakskiy.paymentProvider.repository.PaymentRepository;
-import lombok.RequiredArgsConstructor;
-import lombok.SneakyThrows;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-import java.util.Optional;
+import java.time.LocalDateTime;
 
-import static com.pakskiy.paymentProvider.dto.TransactionStatus.FAILED;
-import static com.pakskiy.paymentProvider.dto.TransactionStatus.IN_PROGRESS;
-import static com.pakskiy.paymentProvider.dto.TransactionType.OUT;
+public interface PayoutService {
+    Mono<PayoutResponseDto> create(PayoutRequestDto request, String token);
 
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class PayoutService {
-    private final MerchantService merchantService;
-    private final PaymentRepository paymentRepository;
-    private final CountryRepository countryRepository;
-    private final CurrencyRepository currencyRepository;
-    private final LanguageRepository languageRepository;
-    private final AccountService accountService;
-    private final ObjectMapper objectMapper;
+    Flux<TransactionEntity> list(LocalDateTime startDate, LocalDateTime endDate);
 
-    @SneakyThrows
-    @Transactional
-    public Mono<PayoutResponseDto> create(PayoutRequestDto request, String token) {
-        return check(request).then(Mono.defer(() -> {
-            Mono<Long> transactionId = merchantService.findByToken(token)
-                    .map(MerchantEntity::getId)
-                    .switchIfEmpty(Mono.error(new RuntimeException("Token not founded")))
-                    .flatMap(accountService::getById)
-                    .switchIfEmpty(Mono.error(new RuntimeException("Account not founded")))
-                    .flatMap(account -> paymentRepository.save(getTransactionEntity(request, account.getMerchantId())))
-                    .map(transactionEntity -> Optional.ofNullable(transactionEntity.getId()).orElse(0L));
-
-            return transactionId.map(res -> {
-                if (res != null && res > 0) {
-                    return PayoutResponseDto.builder().payoutId(res).status(IN_PROGRESS).message("OK").build();
-                }
-                return PayoutResponseDto.builder().status(FAILED).message("PAYOUT_METHOD_NOT_ALLOWED").build();
-            });
-        })).onErrorResume(ex -> {
-            log.error("ERR_CREATE_COMMON {}", ex.getMessage(), ex);
-            return Mono.just(PayoutResponseDto.builder().status(FAILED).message("PAYOUT_METHOD_NOT_ALLOWED").build());
-        });
-    }
-
-    private Mono<Void> check(PayoutRequestDto payoutRequestDto) {
-        return validateCountry(payoutRequestDto.getCustomer().getCountry().toUpperCase())
-                .then(validateCurrency(payoutRequestDto.getCurrency().toUpperCase()))
-                .then(validateLanguage(payoutRequestDto.getLanguage().toUpperCase()));
-    }
-
-    private Mono<Void> validateCountry(String id) {
-        return countryRepository.findById(id)
-                .switchIfEmpty(Mono.error(new RuntimeException("Resource country not found")))
-                .then();
-    }
-
-    private Mono<Void> validateCurrency(String id) {
-        return currencyRepository.findById(id)
-                .switchIfEmpty(Mono.error(new RuntimeException("Resource currency not found")))
-                .then();
-    }
-
-    private Mono<Void> validateLanguage(String id) {
-        return languageRepository.findById(id)
-                .switchIfEmpty(Mono.error(new RuntimeException("Resource language not found")))
-                .then();
-    }
-
-    @SneakyThrows
-    private TransactionEntity getTransactionEntity(PayoutRequestDto request, Long merchantId) {
-        String customerData = objectMapper.writeValueAsString(request.getCustomer());
-        String cardData = objectMapper.writeValueAsString(request.getCardData());
-
-        return TransactionEntity.builder()
-                .accountId(merchantId)
-                .providerTransactionId(request.getProviderTransactionId())
-                .method(request.getPaymentMethod())
-                .amount(-request.getAmount())
-                .currencyId(request.getCurrency().toUpperCase())
-                .createdAt(request.getCreatedAt())
-                .updatedAt(request.getUpdatedAt())
-                .cardData(cardData)//here need parse card data
-                .languageId(request.getLanguage().toUpperCase())
-                .notificationUrl(request.getNotificationUrl())
-                .type(OUT)
-                .customerData(customerData)//here need parse card data
-                .status(IN_PROGRESS)
-                .build();
-    }
+    Mono<TransactionEntity> get(Long transactionId);
 }
