+++ src/main/java/com/pakskiy/paymentProvider/service/TransactionService.java
@@ -2,8 +2,6 @@ package com.pakskiy.paymentProvider.service;
 
 import com.pakskiy.paymentProvider.dto.TransactionRequestDto;
 import com.pakskiy.paymentProvider.dto.TransactionType;
-import com.pakskiy.paymentProvider.dto.payment.PaymentRequestDto;
-import com.pakskiy.paymentProvider.dto.payment.PaymentResponseDto;
 import com.pakskiy.paymentProvider.entity.TransactionEntity;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
@@ -11,9 +9,11 @@ import reactor.core.publisher.Mono;
 import java.time.LocalDateTime;
 
 public interface TransactionService {
-    Mono<PaymentResponseDto> create(TransactionRequestDto request, TransactionType type, String token);
+    Mono<Long> process(TransactionRequestDto request, String token, TransactionType type);
 
     Flux<TransactionEntity> list(LocalDateTime startDate, LocalDateTime endDate, TransactionType type);
 
-    Mono<TransactionEntity> get(Long transactionId);
+    Mono<TransactionEntity> get(Long transactionId, TransactionType type);
+
+    Mono<Void> validate(TransactionRequestDto request);
 }
